# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "COMFACA Crédito API",
        "description": "API para la gestión de solicitudes de crédito de COMFACA",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/solicitudes/{solicitud_id}/generar-pdf": {
            "post": {
                "tags": [
                    "SolicitudPdf"
                ],
                "summary": "Generar PDF de solicitud",
                "description": "Genera el PDF de una solicitud de crédito con soporte para convenios y firmantes.\nEste endpoint reemplaza el antiguo '/api/solicitud-credito/xml' (deprecado).",
                "operationId": "9cad4fd1c897ae7eddbb935153e683bb",
                "parameters": [
                    {
                        "name": "solicitud_id",
                        "in": "path",
                        "description": "ID de la solicitud",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF generado exitosamente"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Solicitud no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/solicitudes/{solicitud_id}/descargar-pdf": {
            "get": {
                "tags": [
                    "SolicitudPdf"
                ],
                "summary": "Descargar PDF de solicitud",
                "description": "Descarga el PDF previamente generado de una solicitud.",
                "operationId": "b6500383b902e4ce3792b2347a615ebc",
                "parameters": [
                    {
                        "name": "solicitud_id",
                        "in": "path",
                        "description": "ID de la solicitud",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF descargado"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "PDF no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/solicitudes/{solicitud_id}/estado-pdf": {
            "get": {
                "tags": [
                    "SolicitudPdf"
                ],
                "summary": "Verificar estado de PDF",
                "description": "Verifica si una solicitud tiene PDF generado y retorna su estado.",
                "operationId": "5b019799d62a4ca8f710de69d776a54d",
                "parameters": [
                    {
                        "name": "solicitud_id",
                        "in": "path",
                        "description": "ID de la solicitud",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estado del PDF"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Solicitud no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/solicitud-pdf/{solicitudId}": {
            "delete": {
                "tags": [
                    "SolicitudPdf"
                ],
                "summary": "Eliminar PDF de solicitud",
                "description": "Eliminar PDF de una solicitud",
                "operationId": "4d6ceec0c7839bab747259324167e2c2",
                "parameters": [
                    {
                        "name": "solicitudId",
                        "in": "path",
                        "description": "ID de la solicitud",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF eliminado"
                    },
                    "404": {
                        "description": "PDF no encontrado"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/solicitud-pdf/estadisticas": {
            "get": {
                "tags": [
                    "SolicitudPdf"
                ],
                "summary": "Obtener estadísticas de PDFs",
                "description": "Obtener estadísticas de PDFs generados",
                "operationId": "b3402633d5a1806cba017a3de927523f",
                "responses": {
                    "200": {
                        "description": "Estadísticas obtenidas"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}